DELETE api will also delete views. bugfix for https://github.com/goccy/bigquery-emulator/issues/190
